+++ conffile.h
@@ -1,5 +1,5 @@
 /*
- * Copyright 2013-2018 Fabian Groffen
+ * Copyright 2013-2019 Fabian Groffen
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -69,6 +69,7 @@ typedef struct _validate {
 typedef struct _cluster {
 	char *name;
 	enum clusttype type;
+	char isdynamic:1;
 	union {
 		chashring *ch;
 		servers *forward;